+++ src/Git_Updater/Theme.php
@@ -267,8 +267,6 @@ class Theme {
 		if ( $schedule_event && ! empty( $themes ) ) {
 			if ( ! $disable_wp_cron && ! $this->is_cron_event_scheduled( 'gu_get_remote_theme' ) ) {
 				wp_schedule_single_event( time(), 'gu_get_remote_theme', [ $themes ] );
-			} else {
-				$this->is_cron_overdue( wp_next_scheduled( 'gu_get_remote_theme', [ $themes ] ) );
 			}
 		}