+++ src/Git_Updater/Traits/GU_Trait.php
@@ -646,8 +646,9 @@ trait GU_Trait {
 	 * @return bool
 	 */
 	public function is_cron_event_scheduled( $hook ) {
-		foreach ( wp_get_ready_cron_jobs() as $event ) {
+		foreach ( wp_get_ready_cron_jobs() as $timestamp => $event ) {
 			if ( $hook === key( $event ) ) {
+				$this->is_cron_overdue( $timestamp );
 				return true;
 			}
 		}