+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 10.4.0 / 2021-07-04 🎆
 * add new WP-Cron task to run concurrently to `wp_version_check` so that Git Updater managed plugins and themes can take advantage of auto updating 🤞
 * added better check to see if background updating cron event is already scheduled
 * add `Ignore()` of certain premium add-ons so not needed in the individual plugins